// web/src/app/shared/models/inputs/TakeDamageInput.ts
export class TakeDamageInput {
  combatId: string;
  creatureId: string;
  physicalDamage: number[];
  arcaneDamage: number[];
  fireDamage: number[];
  iceDamage: number[];
  lightningDamage: number[];
  poisonDamage: number[];
  holyDamage: number[];
  necroticDamage: number[];
  sonicDamage: number[];
}